refactor(net): share utun AF-header framing across endpoints
The macOS utun 4-byte address-family header must be network (big-endian)
byte order — host order makes the kernel see an unknown family and silently
drop the injected packet. That load-bearing encode was duplicated in
DarwinTun (darwin/tun.rs) and the host endpoint (tcpstack utun.rs), so the
byte-order fix lived in two places.

Hoist the encode + AF_HEADER_SIZE into arcbox_tcpstack::utun::utun_af_header
as the single source of truth; DarwinTun now imports it. The readv/writev
wrappers stay local (different fd ownership). Output is byte-identical; the
network-byte-order test moves with the encode.
